LayoutTests/fast/css-grid-layout/non-named-grid-line-get-set-expected.txt

Test that setting and getting grid-template-columns and grid-template-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-template-columns and grid-template-rows set through CSS
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-template-columns') is "first 10px"
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-template-rows') is "first 15px"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-template-columns') is "53% last"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-template-rows') is "27% last"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-template-columns') is "first auto"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-template-rows') is "auto last"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('-webkit-grid-template-columns') is "first minmax(10%, 15px)"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('-webkit-grid-template-rows') is "minmax(20px, 50%) last"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('-webkit-grid-template-columns') is "nav first 10px last"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('-webkit-grid-template-rows') is "nav first 15px last"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('-webkit-grid-template-columns') is "nav first 10% nav 15% last"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('-webkit-grid-template-rows') is "first nav2 25% nav2 75% last"

Test getting and setting grid-template-columns and grid-template-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first 18px"
PASS element.style.webkitGridTemplateColumns is "first 18px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "66px last"
PASS element.style.webkitGridTemplateRows is "66px last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first 55%"
PASS element.style.webkitGridTemplateColumns is "first 55%"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "40% last"
PASS element.style.webkitGridTemplateRows is "40% last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first auto"
PASS element.style.webkitGridTemplateColumns is "first auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "auto last"
PASS element.style.webkitGridTemplateRows is "auto last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first -webkit-min-content"
PASS element.style.webkitGridTemplateColumns is "first -webkit-min-content"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "-webkit-min-content last"
PASS element.style.webkitGridTemplateRows is "-webkit-min-content last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first -webkit-max-content"
PASS element.style.webkitGridTemplateColumns is "first -webkit-max-content"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "-webkit-max-content last"
PASS element.style.webkitGridTemplateRows is "-webkit-max-content last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first minmax(55%, 45px)"
PASS element.style.webkitGridTemplateColumns is "first minmax(55%, 45px)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "minmax(30px, 40%) last"
PASS element.style.webkitGridTemplateRows is "minmax(30px, 40%) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first minmax(220px, -webkit-max-content)"
PASS element.style.webkitGridTemplateColumns is "first minmax(22em, -webkit-max-content)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "minmax(-webkit-max-content, 50px) last"
PASS element.style.webkitGridTemplateRows is "minmax(-webkit-max-content, 5em) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first minmax(220px, -webkit-min-content)"
PASS element.style.webkitGridTemplateColumns is "first minmax(22em, -webkit-min-content)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "minmax(-webkit-min-content, 50px) last"
PASS element.style.webkitGridTemplateRows is "minmax(-webkit-min-content, 5em) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "first minmax(-webkit-min-content, -webkit-max-content)"
PASS element.style.webkitGridTemplateColumns is "first minmax(-webkit-min-content, -webkit-max-content)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "minmax(-webkit-max-content, -webkit-min-content) last"
PASS element.style.webkitGridTemplateRows is "minmax(-webkit-max-content, -webkit-min-content) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "nav first minmax(-webkit-min-content, -webkit-max-content) last"
PASS element.style.webkitGridTemplateColumns is "first nav minmax(-webkit-min-content, -webkit-max-content) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "nav first minmax(-webkit-max-content, -webkit-min-content) last"
PASS element.style.webkitGridTemplateRows is "first nav minmax(-webkit-max-content, -webkit-min-content) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "nav first minmax(-webkit-min-content, -webkit-max-content) nav auto last"
PASS element.style.webkitGridTemplateColumns is "first nav minmax(-webkit-min-content, -webkit-max-content) nav auto last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "first nav2 minmax(-webkit-max-content, -webkit-min-content) nav2 minmax(10px, 15px) last"
PASS element.style.webkitGridTemplateRows is "first nav2 minmax(-webkit-max-content, -webkit-min-content) nav2 minmax(10px, 15px) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "foo bar auto foo auto bar"
PASS element.style.webkitGridTemplateColumns is "foo bar auto foo auto bar"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "foo bar auto foo auto bar"
PASS element.style.webkitGridTemplateRows is "foo bar auto foo auto bar"

Test getting and setting invalid grid-template-columns and grid-template-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-rows') is "none"
PASS successfullyParsed is true

TEST COMPLETE
